chore: enable hashing of various request args
Various request argument structs/enums can now be used as a key in
various *Map and *Set collections like `HashSet`.

The types changed in this commit:
    * struct SessionSetArgs
    * enum Priority
    * struct TorrentAddArgs
    * enum TorrentAction
    * struct TrackerList
    * struct TorrentSetArgs

This change adds a dependency on the `ordered-float` crate to facilitate
`Eq` and `Hash` for the `seed_ratio_limit` (f32) fields in
`SessionSetArgs` and `TorrentSetArgs`.