fix!: don't expose TorrentSetArgs `ids` field
BREAKING CHANGE: This is a breaking change if users were referencing
this field directly.

This field is blindly overwritten by `RpcRequest::torrent_set` so it
should not be exposed to users.

Possibly a better solution would be to leave it exposed (so not a
breaking change) and _somehow_ merge TorrentSetArgs.ids with the `ids`
passed into `RpcRequest::torrent_set`, maybe something like:

    - `torrent_set(..., None)` overrides any TorrentSetArgs.ids
    - any ids defined in TorrentSetArgs are included into ids that are
      passed to `RpcRequest::torrent_set`.

... though, this could have unintended side effects somewhere.